+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # knife-windows Change Log
 
 ## Unreleased changes
-None.
+* [knife-windows #139](https://github.com/opscode/knife-windows/issues/139) Force dev dependency on Chef 11 for test scenarios to avoid Ohai 8 conflict on Ruby 1.9.x
 
 ## Latest release: 0.8.3
 * [knife-windows #131](https://github.com/opscode/knife-windows/issues/108) Issue #131: Windows should be bootstrapped using latest Chef Client version compatible with knife's version just like non-Windows systems